+++ doc/minimal-example.js
@@ -3,12 +3,22 @@ document.addEventListener('deviceready', initializeStore, false);
 
 function initializeStore() {
 
-    // We register a dummy product.
-    CdvPurchase.store.register({
-        id:    "com.example.app.inappid1",
-        alias: "100 coins",
-        type:  CdvPurchase.ProductType.CONSUMABLE
-    });
+    // We register out products.
+    CdvPurchase.store.register([{
+        id:    "my-consumable",
+        type:  CdvPurchase.ProductType.CONSUMABLE,
+        platform: Platform.APPLE_APPSTORE
+    }, {
+        id:    "basic-subscription",
+        type:  CdvPurchase.ProductType.PAID_SUBSCRIPTION,
+        platform: Platform.GOOGLE_PLAY,
+        group: "group-1"
+    }, {
+        id:    "premium-subscription",
+        type:  CdvPurchase.ProductType.PAID_SUBSCRIPTION,
+        platform: Platform.GOOGLE_PLAY,
+        group: "group-1"
+    }]);
 
     // Initialize the store, we just want Google Play and Apple AppStore enabled
     CdvPurchase.store.initialize([CdvPurchase.Platform.GOOGLE_PLAY, CdvPurchase.Platform.APPLE_APPSTORE])
@@ -16,4 +26,3 @@ function initializeStore() {
         console.log("\\o/ STORE READY \\o/");
     });
 }
-