+++ www/store.d.ts
@@ -5647,6 +5647,11 @@ declare namespace CdvPurchase {
                      *
                      * It might be present when the server had to fallback to a backup validation solution. */
                     warning?: string;
+                    /** Date and time the receipt was validated.
+                     *
+                     * It will provide the client with a more reliable clock time
+                     * than the user's device when needed. */
+                    date?: ISODate;
                 };
             }
             type NativeTransaction = ({
@@ -5724,6 +5729,8 @@ declare namespace CdvPurchase {
     class VerifiedReceipt {
         /** @internal */
         className: 'VerifiedReceipt';
+        /** Date and time the receipt was verified */
+        validationDate: Date;
         /** Platform this receipt originated from */
         get platform(): Platform;
         /** Source local receipt used for this validation */